LayoutTests/inspector/sampling-profiler/expression-location-info-expected.txt


== Running test suite: ScriptProfiler.Samples.ExpressionLocation
-- Running test case: Sampling Profiler Expression Location
PASS: Should have seen line 20, column 14.
PASS: Should have seen line 26, column 14.
